refactor: rename localStorage
